requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
timm>=0.9.0
albumentations>=1.3.0
opencv-python>=4.8.0
pillow>=9.5.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Explainability
grad-cam>=1.4.0
shap>=0.42.0

# Uncertainty Quantification
torch-uncertainty>=0.1.0

# MLOps
mlflow>=2.5.0
dvc>=3.0.0

# API and Serving
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Frontend
streamlit>=1.25.0

# Monitoring
prometheus-client>=0.17.0
grafana-api>=1.0.3

# RAG and LLM
langchain>=0.0.300
openai>=1.0.0
faiss-cpu>=1.7.4
chromadb>=0.4.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
